Works in firefox.
Firefox does not support innerText (I suspect this is a Microsoft
thing, and chrome copies Microsoft for compatibility reasons). I
have deferred to Firefox (standards compliance and all that). Again
all this should probably done with jQuery, but since this remains a
learning project, I don't want to do that quite yet.